Revert "development | cambios capitulo 23 y 24 autowired"
This reverts commit 6084a0f2bcd48c6d0c00baf115a97acd0639b7a9.

+++ src/main/java/com/platzimarket/domain/repository/IProductRepository.java
@@ -1,7 +1,6 @@
 package com.platzimarket.domain.repository;
 
 import com.platzimarket.domain.Product;
-import com.platzimarket.persistence.entity.ProductEntity;
 
 import java.util.List;
 import java.util.Optional;
@@ -14,9 +13,9 @@ public interface IProductRepository {
 
   Optional<List<Product>> getScarseProducts(int quantity);
 
-  Optional<Product> getProduct(long productId);
+  Optional<Product> getProduct(int productId);
 
   Product save(Product product);
 
-  void delete(long productId);
+  void delete(int productId);
 }
